Fixes #124

+++ kikaha-maven-plugin/source/META-INF/defaults/bin/kikaha.sh
@@ -1,5 +1,5 @@
 #!/bin/sh
-cd "`dirname $0`/.."
+cd "`dirname $(readlink -f $0)`/.."
 
 # CONFIGURABLE VARIABLES
 LIBDIR="./lib"
@@ -10,11 +10,44 @@ JAVA_OPTS=
 MAIN_CLASS=kikaha.core.Main
 OUTPUTFILE=server.stdout.log
 PIDFILE=server.pid
+NULL=/dev/null
+
+print_logo(){
+if [ ! "$QUIET" = "true" ]; then
+cat <<EOF
+      $(yellow " __  _  ____  __  _   ____  __ __   ____ ")
+      $(yellow "|  |/ ]|    ||  |/ ] /    ||  |  | /    |")
+      $(yellow "|  ' /  |  | |  ' / |  o  ||  |  ||  o  |")
+      $(yellow "|    \  |  | |    \ |     ||  _  ||     |")
+      $(yellow "|     | |  | |     ||  _  ||  |  ||  _  |")
+      $(yellow "|  .  | |  | |  .  ||  |  ||  |  ||  |  |")
+      $(yellow "|__|\_||____||__|\_||__|__||__|__||__|__|")
+ $(grape 'an absurdly fast web server designed for microservices')
+
+EOF
+fi
+}
+
+yellow(){
+        echo -n "\033[1;33m$@\033[m"
+}
+
+red(){
+        echo -n "\033[1;31m$@\033[m"
+}
+
+grape(){
+        echo -n "\033[1;35m$@\033[m"
+}
 
 start_server(){
-	echo "Starting server in background..."
-	nohup ${JAVA} ${JAVA_OPTS} -classpath "${CLASSPATH}" ${MAIN_CLASS} > $OUTPUTFILE &
-	echo "$!" > $PIDFILE
+	if [ -e $PIDFILE ]; then
+		red "[WARN] Server already running"
+	else
+		yellow "[INFO] Starting server in background..."
+		nohup ${JAVA} ${JAVA_OPTS} -classpath "${CLASSPATH}" ${MAIN_CLASS} > $NULL 2> $NULL &
+		echo "$!" > $PIDFILE
+	fi
 }
 
 stop_server(){
@@ -22,11 +55,10 @@ stop_server(){
 		PID=`cat $PIDFILE`
 		FOUNDPROCS=`ps aux | grep $PID | head -n 1 | grep java`
 		if [ ! "$FOUNDPROCS" = "" ]; then
-			kill $PID && echo "kill term sent."
-			rm -f $PIDFILE
+			kill $PID && echo "kill term sent." && rm -f $PIDFILE
 		fi
 	else
-		echo "Server not running"
+		red "[WARN] Server not running"
 		exit 1
 	fi
 }
@@ -37,6 +69,7 @@ show_help(){
 	echo "Available commands:"
 	echo " - start:	starts the server in background"
 	echo " - stop:	stops the server that is running in background"
+	echo " - restart: restarts the server in background"
 	echo " - debug:	run the server in foreground"
 	echo " - help:	shows this help message"
 	echo
@@ -51,9 +84,11 @@ fi
 # MAIN
 CLASSPATH=".:${LIBDIR}/*"
 
+print_logo
 case "$1" in
 	"stop" ) stop_server ;;
 	"debug" ) ${JAVA} ${JAVA_OPTS} -classpath "${CLASSPATH}" ${MAIN_CLASS} ;;
 	"start" ) start_server ;;
+	"restart" ) stop_server && start_server ;;
 	"help" | * ) show_help ;;
 esac
